feat: wip
What's the best way to deal with it? selling and buying with price 0 or with the mean price per share, or with the price at market time? The first two ways are giving wrong results in the average unit price calculations, and the last one was not tested yet. For now, we're going to just chnage the ticker name using a mapper file.